Fix mobile mixer values

+++ FamiStudio/Source/App/Mobile/Grid.cs
@@ -312,7 +312,7 @@ namespace FamiStudio
                 // Scale value based on the format, then rescale and store as int.
                 var scale = Utils.ParseFloatWithTrailingGarbage(slider.Format(1));
                 var name  = GetControlLabel(sender) ?? string.Empty;
-                var val   = Math.Round(slider.Value * scale, 2);
+                var val   = Math.Round((int)slider.Value * scale, 2);
 
                 Platform.EditTextAsync(name, val.ToString("0.#", CultureInfo.InvariantCulture), (s) =>
                 {
